Backport PRs 155, 157 and 158 (#159)
* Serialization long type incorrect for Linux 32-bit, Add more tests for ReliableTopic for increasing coverage (#155)

* Added more tests for ReliableTopic.

* A fix is done at ReliableTopicExecutor.

* Changed long to int64_t for serialization.

* Add Reliable Topic Code Sample (#157)

* Added code examples for reliable topic usage.

* Windows portability fix: Changed static topic prefix string to a member to allow windows dllexport work with it.

* Ringbuffer and ReliableTopic documentation.

+++ hazelcast/include/hazelcast/client/serialization/pimpl/DataInput.h
@@ -15,10 +15,6 @@
  */
 //
 // Created by sancar koyunlu on 8/7/13.
-
-
-
-
 #ifndef HAZELCAST_DataInput
 #define HAZELCAST_DataInput
 
@@ -26,10 +22,13 @@
 #include "hazelcast/util/ByteBuffer.h"
 #include "hazelcast/util/Bits.h"
 #include "hazelcast/client/exception/HazelcastSerializationException.h"
+
+#include <boost/smart_ptr/shared_ptr.hpp>
+
 #include <vector>
 #include <string>
 #include <memory>
-#include <boost/smart_ptr/shared_ptr.hpp>
+#include <stdint.h>
 
 #if  defined(WIN32) || defined(_WIN32) || defined(WIN64) || defined(_WIN64)
 #pragma warning(push)
@@ -66,7 +65,7 @@ namespace hazelcast {
 
                     int readInt();
 
-                    long long readLong();
+                    int64_t readLong();
 
                     float readFloat();
